read data from strings instead of files in tests
