--- include/generators/order_polytope/instances/bayesiannetwork_andes_008_0.txt
@@ -1,8 +1,0 @@
-0 0 0 0 0 0 0 0
-1 0 0 0 0 0 0 0
-0 1 0 0 0 0 0 0
-0 0 0 0 0 0 0 0
-0 0 1 1 0 0 0 0
-0 0 0 0 1 0 0 0
-0 0 0 0 0 0 0 0
-0 0 0 0 1 1 1 0
