--- include/generators/order_polytope/instances/bipartite_0.5_008_0.txt
@@ -1,8 +1,0 @@
-0 0 0 0 1 0 1 0
-0 0 0 0 1 0 0 0
-0 0 0 0 1 1 0 1
-0 0 0 0 1 0 1 0
-0 0 0 0 0 0 0 0
-0 0 0 0 0 0 0 0
-0 0 0 0 0 0 0 0
-0 0 0 0 0 0 0 0
